requirements.txt
numpy == 1.20.1
pandas == 1.2.4
matplotlib == 3.3.4
seaborn  == 0.10.1
scikit-learn == 1.0.2
imbalanced-learn == 0.9.0
xgboost == 1.5.2
shap == 0.40.0
joblib == 1.0.1
Flask == 1.1.2
requests == 2.27.1
streamlit == 1.5.0
pillow = 8.2.0